Split modules for server view
