+++ src/deepLinks/main.ts
@@ -7,7 +7,7 @@ import { resolveServerUrl } from '../servers/main';
 import { select, dispatch } from '../store';
 import { askForServerAddition, warnAboutInvalidServerUrl } from '../ui/main/dialogs';
 import { getRootWindow } from '../ui/main/rootWindow';
-import { getWebContentsByServerUrl } from '../ui/main/webviews';
+import { getWebContentsByServerUrl } from '../ui/main/serverView';
 import {
   DEEP_LINKS_SERVER_FOCUSED,
   DEEP_LINKS_SERVER_ADDED,
