+++ src/main.ts
@@ -22,9 +22,9 @@ import {
   showRootWindow,
   exportLocalStorage,
 } from './ui/main/rootWindow';
+import { attachGuestWebContentsEvents } from './ui/main/serverView';
 import touchBar from './ui/main/touchBar';
 import trayIcon from './ui/main/trayIcon';
-import { attachGuestWebContentsEvents } from './ui/main/webviews';
 import { setupUpdates } from './updates/main';
 import { setupPowerMonitor } from './userPresence/main';
